feat(green): allow both Optional and required flags to have defaults, changing semantics slightly
