+++ experiments/simple2/tests/testData.ts
@@ -24,9 +24,6 @@ export function makeDashDashCase(
 ): ArgsExample {
   if (insertBefore >= eg.raw.length) { // dashdash at end causes error
     const raw = [...eg.raw, "--"];
-    // if (insertBefore < 4) {
-    //   console.debug("failing on raw=", raw, insertBefore, eg.raw);
-    // }
     return { raw, parsed: DashedWithoutArgs };
   }
   const args = eg.raw.slice(0, insertBefore);
